fix: rename src/features.{c,h} to curry_features.{c,h}
`src/features.h` shadowed glibc's own foundational system header of
the exact same name (`<features.h>`, which defines `__GLIBC_USE` and
friends and is transitively included by nearly everything, including
Boehm GC's own headers). On Linux, with `src/` in the include search
path ahead of the system include dirs, `#include <features.h>` from
gc/gc_config_macros.h resolved to curry's own file instead of glibc's,
breaking the build everywhere that transitively includes <gc/gc.h> --
which is nearly every translation unit. Never surfaced locally
(macOS/BSD libc has no header by this name), only caught by CI's
Ubuntu build.
