+++ src/curry_features.c
@@ -1,4 +1,4 @@
-#include "features.h"
+#include "curry_features.h"
 #include "object.h"
 #include "symbol.h"
 #include "builtins.h"
